+++ resources/views/welcome.blade.php
@@ -274,6 +274,6 @@
         @if (Route::has('login'))
             <div class="h-14.5 hidden lg:block"></div>
         @endif
-        @livewireScripts
+        @livewireScriptConfig
     </body>
 </html>